Make dracputester build on non-Linux hosts
