+++ test/image/dracputester/affinity_linux.go
@@ -1,0 +1,41 @@
+//go:build linux
+
+/*
+Copyright 2025 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+import (
+	"os"
+
+	"github.com/go-logr/logr"
+	"github.com/kubernetes-sigs/dra-driver-cpu/pkg/cpuinfo"
+	"golang.org/x/sys/unix"
+	"k8s.io/utils/cpuset"
+)
+
+func getAffinity(logger logr.Logger) (cpuset.CPUSet, error) {
+	var unixCS unix.CPUSet
+	err := unix.SchedGetaffinity(os.Getpid(), &unixCS)
+	if err != nil {
+		return cpuset.New(), err
+	}
+	maxCPUID := 0
+	if topo, err := cpuinfo.NewSystemCPUInfo().GetCPUTopology(logger); err == nil {
+		maxCPUID = affinityScanBoundFromTopology(topo)
+	}
+	return affinityFromMask(&unixCS, maxCPUID), nil
+}
