+++ test/image/dracputester/affinity_other.go
@@ -1,0 +1,32 @@
+//go:build !linux
+
+/*
+Copyright 2025 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+import (
+	"fmt"
+	"runtime"
+
+	"github.com/go-logr/logr"
+	"k8s.io/utils/cpuset"
+)
+
+func getAffinity(logger logr.Logger) (cpuset.CPUSet, error) {
+	_ = logger
+	return cpuset.New(), fmt.Errorf("CPU affinity detection requires linux, current platform is %s", runtime.GOOS)
+}
